+++ controllers/contacts/index.js
@@ -1,0 +1,13 @@
+const getAll = require("./getAll");
+const getById = require("./getById");
+const post = require("./post");
+const removeById = require("./removeById");
+const put = require("./put");
+
+module.exports = {
+  getAll,
+  getById,
+  post,
+  removeById,
+  put,
+};
