--- controllers/index.js
@@ -1,5 +1,0 @@
-const contacts = require("./contacts");
-
-module.exports = {
-  contacts,
-};
