+++ middlewares/index.js
@@ -1,3 +1,4 @@
 const controllerWrapper = require("./controllerWrapper");
+const validation = require("./validation");
 
-module.exports = { controllerWrapper };
+module.exports = { controllerWrapper, validation };
