+++ model/contacts/readContacts.js
@@ -1,0 +1,8 @@
+const fs = require("fs/promises");
+
+const readContacts = async (path) => {
+  const contacts = await fs.readFile(path);
+  return JSON.parse(contacts);
+};
+
+module.exports = { readContacts };
